% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column_unqdetecid.r
\name{add_column_unqdetecid}
\alias{add_column_unqdetecid}
\title{Add Unique Detection ID}
\usage{
add_column_unqdetecid(df)
}
\arguments{
\item{df}{A dataframe with no duplicate rows}
}
\description{
Adds an explicit incremental integer index column to a dataframe for compatibility with other functionality
}
\examples{
df <- add_column_unqdetecid(df)
filtered_df <- filter_detections(add_column_unqdetecid('my-unindexed-detections.csv'))
}
\keyword{add}
\keyword{id}
\keyword{unique}
