% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cohort_script.r
\name{cohorts}
\alias{cohorts}
\title{Cohort Identification}
\usage{
cohorts(comp_df, time_interval = 60)
}
\arguments{
\item{comp_df}{A compressed detection dataframe}
}
\description{
This function identifies animals who are co-detected on stations within a time threshold (in mins)
}
\examples{
cohorts(compress_detections(read.csv('my-detection-file.csv')))

}
\keyword{acoustic}
\keyword{compress}
\keyword{telemetry}
\keyword{temporal}
