% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_detections.r
\name{compress_detections}
\alias{compress_detections}
\title{Compressed Detection Dataframe}
\usage{
compress_detections(detection_df)
}
\arguments{
\item{detection_df}{A dataframe read from a CSV file of detections}
}
\description{
This function compresses concurrent detections of a tag at a single station into an aggregate w/ simple statistics
}
\examples{
compress_detections(read.csv('my-detection-file.csv'))
}
\keyword{acoustic}
\keyword{compress}
\keyword{telemetry}
\keyword{temporal}
