% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_detections.r
\name{filter_detections}
\alias{filter_detections}
\title{Temporal Filter Tool}
\usage{
filter_detections(detection_df, suspect_df = NULL, detection_radius = NULL,
  min_time_buffer = 60, distance_matrix = F)
}
\arguments{
\item{detection_df}{A dataframe read from a CSV file of detections}

\item{suspect_df}{A dataframe of suspect detections from a previous or different filter run (not yet implemented)}

\item{detection_radius}{A detection radius. (not used)}

\item{min_time_buffer}{How far from other detections a detection must be to be filtered (in minutes)}

\item{distance_matrix}{Whether or not to generate a distance matrix from the supplied station locations.}
}
\description{
This function filters out acoustic detections that are isolated by more than N minutes on either side
}
\examples{
filter_detections(read.csv('my-detection-file.csv'))
filter_detections(read.csv('my-detection-file.csv'), min_time_buffer=30, distance_matrix=T)
}
\keyword{acoustic}
\keyword{filter}
\keyword{telemetry}
\keyword{temporal}
